// utils/types.ts
export type Tool = | {
    type: "rect";
    id: string;
    x: number;
    y: number;
    width: number;
    height: number;
    color: string;
    bgColor: string;
    strokeWidth: number;
    strokeStyle: string;
    rotation?: number;
  }| {
    type: "circle";
    id: string;
    x: number;
    y: number;
    endX: number;
    endY: number;
    width: number;
    height: number;
    centerX: number;
    centerY: number;
    radius: number;
    color: string;
    bgColor: string;
    strokeWidth: number;
    strokeStyle: string;
    rotation?: number;
  }| {
    type: "pencil";
    id: string;
    path: { x: number; y: number }[];
    color: string;
    strokeWidth: number;
    strokeStyle: string;
  }| {
    type: "diamond";
    id: string;
    x: number;
    y: number;
    width: number;
    height: number;
    color: string;
    bgColor: string;
    strokeWidth: number;
    strokeStyle: string;
    rotation?: number;
  }| {
    type: "arrow";
    id: string;
    startX: number;
    startY: number;
    endX: number;
    endY: number;
    color: string;
    strokeWidth: number;
    strokeStyle: string;
    rotation?: number;
  }| {
    type: "line";
    id: string;
    startX: number;
    startY: number;
    endX: number;
    endY: number;
    color: string;
    strokeWidth: number;
    strokeStyle: string;
    rotation?: number;
  }| {
    type: "text";
    id: string;
    x: number;
    y: number;
    text: string;
    size: number;
    color: string;
    bgColor: string;
    strokeWidth: number;
    strokeStyle: string;
    rotation?: number;
  }| {
    type: "eraser";
    id: string;
    path: { x: number; y: number }[];
    strokeWidth: number;
  }| {
    type: "select";
    id: string;
    x: number;
    y: number;
    width: number;
    height: number;
  };

  export type ToolType = Tool['type'];